backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0
boto3>=1.34.0
Pillow>=10.0.0
